chore: move images
