+++ src/assets/images/fones.png
