+++ src/assets/images/mapa_icon.png
